+++ src/main/java/com/iexec/sms/tee/challenge/TeeChallenge.java
@@ -16,16 +16,18 @@
 
 package com.iexec.sms.tee.challenge;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
 
-@Data
+@Entity
 @Getter
 @Builder
 @NoArgsConstructor
-@Entity
 @AllArgsConstructor
 public class TeeChallenge {