+++ src/test/java/com/iexec/sms/secret/web2/Web2SecretServiceTests.java
@@ -16,14 +16,20 @@
 
 package com.iexec.sms.secret.web2;
 
+import ch.qos.logback.classic.Logger;
+import com.iexec.sms.MemoryLogAppender;
 import com.iexec.sms.encryption.EncryptionService;
 import com.iexec.sms.secret.CacheSecretService;
 import com.iexec.sms.secret.MeasuredSecretService;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
+import org.junit.jupiter.api.TestInstance;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import java.util.Optional;
 
@@ -31,13 +37,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
+@DataJpaTest
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class Web2SecretServiceTests {
     private static final String OWNER_ADDRESS = "ownerAddress";
     private static final String SECRET_ADDRESS = "secretAddress";
     private static final String PLAIN_SECRET_VALUE = "plainSecretValue";
     private static final String ENCRYPTED_SECRET_VALUE = "encryptedSecretValue";
 
-    @Mock
+    @Autowired
     private Web2SecretRepository web2SecretRepository;
 
     @Mock
@@ -46,25 +54,34 @@ class Web2SecretServiceTests {
     @Mock
     private MeasuredSecretService measuredSecretService;
 
-    @Mock
     private CacheSecretService<Web2SecretHeader> web2CacheSecretService;
 
-    @InjectMocks
     private Web2SecretService web2SecretService;
 
+    private static MemoryLogAppender memoryLogAppender;
+
+    @BeforeAll
+    void initLog() {
+        Logger logger = (Logger) LoggerFactory.getLogger("com.iexec.sms.secret");
+        memoryLogAppender = (MemoryLogAppender) logger.getAppender("MEM");
+    }
 
     @BeforeEach
     void beforeEach() {
         MockitoAnnotations.openMocks(this);
+        memoryLogAppender.reset();
+        web2CacheSecretService = new CacheSecretService<>();
+        web2SecretRepository.deleteAll();
+        web2SecretService = new Web2SecretService(web2SecretRepository, encryptionService, measuredSecretService,web2CacheSecretService);
     }
+
 
     // region getSecret
     @Test
     void shouldGetDecryptedValue() {
         final Web2Secret encryptedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
+        web2SecretRepository.save(encryptedSecret);
 
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.of(encryptedSecret));
         when(encryptionService.decrypt(ENCRYPTED_SECRET_VALUE))
                 .thenReturn(PLAIN_SECRET_VALUE);
 
@@ -77,9 +94,7 @@ class Web2SecretServiceTests {
     @Test
     void shouldGetEncryptedSecret() {
         final Web2Secret encryptedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
-
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.of(encryptedSecret));
+        web2SecretRepository.save(encryptedSecret);
 
         final Optional<Web2Secret> result = web2SecretService.getSecret(OWNER_ADDRESS, SECRET_ADDRESS);
         assertThat(result).isNotEmpty();
@@ -92,23 +107,13 @@ class Web2SecretServiceTests {
 
     @Test
     void shouldGetEmptySecretIfSecretNotPresent() {
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.empty());
-
         assertThat(web2SecretService.getSecret(OWNER_ADDRESS, SECRET_ADDRESS)).isEmpty();
-        verify(web2SecretRepository, times(1))
-                .findById(any(Web2SecretHeader.class));
         verifyNoInteractions(encryptionService);
     }
 
     @Test
     void shouldGetEmptyDecryptedValueIfSecretNotPresent() {
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.empty());
-
         assertThat(web2SecretService.getDecryptedValue(OWNER_ADDRESS, SECRET_ADDRESS)).isEmpty();
-        verify(web2SecretRepository, times(1))
-                .findById(any(Web2SecretHeader.class));
         verifyNoInteractions(encryptionService);
     }
     // endregion
@@ -118,39 +123,47 @@ class Web2SecretServiceTests {
     @Test
     void shouldGetSecretExistFromDBAndPutInCache() {
         final Web2Secret encryptedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
+        web2SecretRepository.save(encryptedSecret);
 
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.of(encryptedSecret));
-        when(web2CacheSecretService.lookSecretExistenceInCache(any(Web2SecretHeader.class))).thenReturn(null);
-
-        final boolean isSecretPresent = web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
+        final boolean resultFirstCall = web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
 
         assertAll(
-                () -> assertTrue(isSecretPresent),
-                () -> verify(web2SecretRepository, times(1)).findById(any())
+                () -> assertTrue(resultFirstCall),
+                () -> assertTrue(memoryLogAppender.contains("Search secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Secret existence was not found in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Put secret existence in cache"))
         );
     }
 
     @Test
     void shouldGetSecretExistFromCache() {
         final Web2Secret encryptedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
-        when(web2CacheSecretService.lookSecretExistenceInCache(any(Web2SecretHeader.class))).thenReturn(true);
+        web2SecretRepository.save(encryptedSecret);
 
-        final boolean isSecretPresent = web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
+        web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
+        memoryLogAppender.reset();
+        final boolean resultSecondCall = web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
         assertAll(
-                () -> assertTrue(isSecretPresent),
-                () -> verify(web2SecretRepository, times(0)).findById(any())
+                () -> assertTrue(resultSecondCall),
+                () -> assertTrue(memoryLogAppender.doesNotContains("Put secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Search secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Secret existence was found in cache")),
+                () -> assertTrue(memoryLogAppender.contains("exist:true"))
 
         );
     }
 
     @Test
     void shouldGetSecretNotExistFromCache() {
-        when(web2CacheSecretService.lookSecretExistenceInCache(any(Web2SecretHeader.class))).thenReturn(false);
-        boolean isSecretPresent = web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
+        web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
+        memoryLogAppender.reset();
+        boolean resultSecondCall = web2SecretService.isSecretPresent(OWNER_ADDRESS, SECRET_ADDRESS);
         assertAll(
-                () -> assertFalse(isSecretPresent),
-                () -> verify(web2SecretRepository, times(0)).findById(any())
+                () -> assertFalse(resultSecondCall),
+                () -> assertTrue(memoryLogAppender.doesNotContains("Put secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Search secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Secret existence was found in cache")),
+                () -> assertTrue(memoryLogAppender.contains("exist:false"))
         );
     }
     // endregion
@@ -158,28 +171,21 @@ class Web2SecretServiceTests {
     // region addSecret
     @Test
     void shouldAddSecret() throws SecretAlreadyExistsException {
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.empty());
         when(encryptionService.encrypt(PLAIN_SECRET_VALUE)).thenReturn(ENCRYPTED_SECRET_VALUE);
-        when(web2SecretRepository.save(any())).thenReturn(new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE));
 
         final Web2Secret newSecret = web2SecretService.addSecret(OWNER_ADDRESS, SECRET_ADDRESS, PLAIN_SECRET_VALUE);
         assertAll(
                 () -> assertThat(newSecret).extracting(Web2Secret::getHeader).usingRecursiveComparison().isEqualTo(new Web2SecretHeader(OWNER_ADDRESS, SECRET_ADDRESS)),
                 () -> assertThat(newSecret).extracting(Web2Secret::getValue).isEqualTo(ENCRYPTED_SECRET_VALUE),
                 () -> verify(measuredSecretService).newlyAddedSecret(),
-
-                () -> verify(encryptionService).encrypt(any()),
-                () -> verify(web2SecretRepository).save(any())
+                () -> verify(encryptionService).encrypt(any())
         );
     }
 
     @Test
     void shouldNotAddSecretIfPresent() {
         final Web2Secret secret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.of(secret));
-        when(web2CacheSecretService.lookSecretExistenceInCache(any(Web2SecretHeader.class))).thenReturn(null);
+        web2SecretRepository.save(secret);
 
         final SecretAlreadyExistsException exception = assertThrows(SecretAlreadyExistsException.class,
                 () -> web2SecretService.addSecret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE));
@@ -189,7 +195,7 @@ class Web2SecretServiceTests {
                 () -> verify(measuredSecretService, times(0)).newlyAddedSecret(),
 
                 () -> verify(encryptionService, never()).encrypt(PLAIN_SECRET_VALUE),
-                () -> verify(web2SecretRepository, never()).save(any())
+                () -> assertThat(web2SecretRepository.count()).isOne()
         );
     }
     // endregion
@@ -200,43 +206,33 @@ class Web2SecretServiceTests {
         final Web2Secret encryptedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
         final String newSecretValue = "newSecretValue";
         final String newEncryptedSecretValue = "newEncryptedSecretValue";
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.of(encryptedSecret));
+        web2SecretRepository.save(encryptedSecret);
         when(encryptionService.encrypt(newSecretValue))
                 .thenReturn(newEncryptedSecretValue);
-        final Web2Secret savedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, newEncryptedSecretValue);
-        when(web2SecretRepository.save(any()))
-                .thenReturn(savedSecret);
 
         final Web2Secret newSecret = web2SecretService.updateSecret(OWNER_ADDRESS, SECRET_ADDRESS, newSecretValue);
         assertAll(
                 () -> assertThat(newSecret).extracting(Web2Secret::getHeader).usingRecursiveComparison().isEqualTo(new Web2SecretHeader(OWNER_ADDRESS, SECRET_ADDRESS)),
                 () -> assertThat(newSecret).extracting(Web2Secret::getValue).isEqualTo(newEncryptedSecretValue),
-
-                () -> verify(web2SecretRepository, never()).save(encryptedSecret),   // Current object should not be updated
-                () -> verify(web2SecretRepository, times(1)).save(any())    // A new object should be created with the same ID
+                () -> assertThat(web2SecretRepository.count()).isOne()    // A new object should be created with the same ID
         );
     }
 
     @Test
     void shouldNotUpdateSecretIfMissing() {
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.empty());
-
         final NotAnExistingSecretException exception = assertThrows(NotAnExistingSecretException.class,
                 () -> web2SecretService.updateSecret(OWNER_ADDRESS, SECRET_ADDRESS, PLAIN_SECRET_VALUE));
         assertAll(
                 () -> assertEquals(OWNER_ADDRESS, exception.getOwnerAddress()),
                 () -> assertEquals(SECRET_ADDRESS, exception.getSecretAddress()),
-                () -> verify(web2SecretRepository, never()).save(any())
+                () -> assertThat(web2SecretRepository.count()).isZero()
         );
     }
 
     @Test
     void shouldNotUpdateSecretIfSameValue() {
         final Web2Secret encryptedSecret = new Web2Secret(OWNER_ADDRESS, SECRET_ADDRESS, ENCRYPTED_SECRET_VALUE);
-        when(web2SecretRepository.findById(any(Web2SecretHeader.class)))
-                .thenReturn(Optional.of(encryptedSecret));
+        web2SecretRepository.save(encryptedSecret);
         when(encryptionService.encrypt(PLAIN_SECRET_VALUE))
                 .thenReturn(ENCRYPTED_SECRET_VALUE);
 
@@ -245,7 +241,7 @@ class Web2SecretServiceTests {
         assertAll(
                 () -> assertEquals(OWNER_ADDRESS, exception.getOwnerAddress()),
                 () -> assertEquals(SECRET_ADDRESS, exception.getSecretAddress()),
-                () -> verify(web2SecretRepository, never()).save(any())
+                () -> assertThat(web2SecretRepository.count()).isOne()
         );
     }
     // endregion