+++ src/test/java/com/iexec/sms/secret/web3/Web3SecretServiceTests.java
@@ -16,14 +16,20 @@
 
 package com.iexec.sms.secret.web3;
 
+import ch.qos.logback.classic.Logger;
+import com.iexec.sms.MemoryLogAppender;
 import com.iexec.sms.encryption.EncryptionService;
 import com.iexec.sms.secret.CacheSecretService;
 import com.iexec.sms.secret.MeasuredSecretService;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
+import org.junit.jupiter.api.TestInstance;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import java.util.Optional;
 
@@ -32,54 +38,65 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
+@DataJpaTest
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class Web3SecretServiceTests {
     String secretAddress = "secretAddress".toLowerCase();
     String plainSecretValue = "plainSecretValue";
     String encryptedSecretValue = "encryptedSecretValue";
 
+    @Autowired
+    private Web3SecretRepository web3SecretRepository;
+
     @Mock
     private EncryptionService encryptionService;
-    @Mock
-    private Web3SecretRepository web3SecretRepository;
+
     @Mock
     private MeasuredSecretService measuredSecretService;
 
-    @Mock
     private CacheSecretService<Web3SecretHeader> web3CacheSecretService;
 
-    @InjectMocks
     private Web3SecretService web3SecretService;
+
+    private static MemoryLogAppender memoryLogAppender;
+
+    @BeforeAll
+    void initLog() {
+        Logger logger = (Logger) LoggerFactory.getLogger("com.iexec.sms.secret");
+        memoryLogAppender = (MemoryLogAppender) logger.getAppender("MEM");
+    }
 
     @BeforeEach
     void beforeEach() {
         MockitoAnnotations.openMocks(this);
+        web3CacheSecretService = new CacheSecretService<>();
+        memoryLogAppender.reset();
+        web3SecretRepository.deleteAll();
+        web3SecretService = new Web3SecretService(web3SecretRepository, encryptionService, measuredSecretService,web3CacheSecretService);
     }
 
     // region addSecret
     @Test
     void shouldNotAddSecretIfPresent() {
         Web3Secret web3Secret = new Web3Secret(secretAddress, encryptedSecretValue);
-        when(web3CacheSecretService.lookSecretExistenceInCache(any(Web3SecretHeader.class))).thenReturn(null);
-
-        when(web3SecretRepository.findById(any(Web3SecretHeader.class))).thenReturn(Optional.of(web3Secret));
+        web3SecretRepository.save(web3Secret);
         assertThat(web3SecretService.addSecret(secretAddress, plainSecretValue)).isFalse();
         verify(measuredSecretService, times(0)).newlyAddedSecret();
         verifyNoInteractions(encryptionService);
-        verify(web3SecretRepository, never()).save(any());
+        assertThat(web3SecretRepository.count()).isOne();
     }
 
     @Test
     void shouldAddSecret() {
-        when(web3SecretRepository.findById(any(Web3SecretHeader.class))).thenReturn(Optional.empty());
         when(encryptionService.encrypt(plainSecretValue)).thenReturn(encryptedSecretValue);
-        when(web3SecretRepository.save(any())).thenReturn(new Web3Secret(secretAddress, encryptedSecretValue));
 
         final boolean success = web3SecretService.addSecret(secretAddress, plainSecretValue);
         assertAll(
                 () -> assertTrue(success),
                 () -> verify(measuredSecretService).newlyAddedSecret(),
                 () -> verify(encryptionService).encrypt(any()),
-                () -> verify(web3SecretRepository).save(any())
+                () -> assertThat(web3SecretRepository.count()).isOne(),
+                () -> assertTrue(memoryLogAppender.contains("Put secret existence in cache"))
         );
     }
     // endregion
@@ -89,35 +106,45 @@ class Web3SecretServiceTests {
     @Test
     void shouldGetSecretExistFromDBAndPutInCache() {
         Web3Secret web3Secret = new Web3Secret(secretAddress, encryptedSecretValue);
-        when(web3CacheSecretService.lookSecretExistenceInCache(any(Web3SecretHeader.class))).thenReturn(null);
-        when(web3SecretRepository.findById(any(Web3SecretHeader.class))).thenReturn(Optional.of(web3Secret));
+        web3SecretRepository.save(web3Secret);
 
-        final boolean isSecretPresent = web3SecretService.isSecretPresent(secretAddress);
+        final boolean resultFirstCall = web3SecretService.isSecretPresent(secretAddress);
         assertAll(
-                () -> assertTrue(isSecretPresent),
-                () -> verify(web3SecretRepository, times(1)).findById(any())
+                () -> assertTrue(resultFirstCall),
+                () -> assertTrue(memoryLogAppender.contains("Search secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Secret existence was not found in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Put secret existence in cache"))
         );
     }
 
     @Test
     void shouldGetSecretExistFromCache() {
-        when(web3CacheSecretService.lookSecretExistenceInCache(any(Web3SecretHeader.class))).thenReturn(true);
+        Web3Secret web3Secret = new Web3Secret(secretAddress, encryptedSecretValue);
+        web3SecretRepository.save(web3Secret);
 
-        boolean isSecretPresent = web3SecretService.isSecretPresent(secretAddress);
+        web3SecretService.isSecretPresent(secretAddress);
+        memoryLogAppender.reset();
+        boolean resultSecondCall = web3SecretService.isSecretPresent(secretAddress);
         assertAll(
-                () -> assertTrue(isSecretPresent),
-                () -> verify(web3SecretRepository, times(0)).findById(any())
+                () -> assertTrue(resultSecondCall),
+                () -> assertTrue(memoryLogAppender.doesNotContains("Put secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Search secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Secret existence was found in cache")),
+                () -> assertTrue(memoryLogAppender.contains("exist:true"))
         );
     }
 
     @Test
     void shouldGetSecretNotExistFromCache() {
-        when(web3CacheSecretService.lookSecretExistenceInCache(any(Web3SecretHeader.class))).thenReturn(false);
-
-        boolean isSecretPresent = web3SecretService.isSecretPresent(secretAddress);
+        web3SecretService.isSecretPresent(secretAddress);
+        memoryLogAppender.reset();
+        boolean resultSecondCall = web3SecretService.isSecretPresent(secretAddress);
         assertAll(
-                () -> assertFalse(isSecretPresent),
-                () -> verify(web3SecretRepository, times(0)).findById(any())
+                () -> assertFalse(resultSecondCall),
+                () -> assertTrue(memoryLogAppender.doesNotContains("Put secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Search secret existence in cache")),
+                () -> assertTrue(memoryLogAppender.contains("Secret existence was found in cache")),
+                () -> assertTrue(memoryLogAppender.contains("exist:false"))
         );
     }
     // endregion
@@ -126,23 +153,18 @@ class Web3SecretServiceTests {
     @Test
     void shouldGetDecryptedValue() {
         Web3Secret encryptedSecret = new Web3Secret(secretAddress, encryptedSecretValue);
-        when(web3SecretRepository.findById(any(Web3SecretHeader.class))).thenReturn(Optional.of(encryptedSecret));
+        web3SecretRepository.save(encryptedSecret);
         when(encryptionService.decrypt(encryptedSecretValue)).thenReturn(plainSecretValue);
 
         Optional<String> result = web3SecretService.getDecryptedValue(secretAddress);
-        assertThat(result)
-                .isPresent()
-                .get().isEqualTo(plainSecretValue);
+        assertThat(result).contains(plainSecretValue);
 
-        verify(web3SecretRepository).findById(any(Web3SecretHeader.class));
         verify(encryptionService).decrypt(any());
     }
 
     @Test
     void shouldGetEmptyValueIfSecretNotPresent() {
-        when(web3SecretRepository.findById(any(Web3SecretHeader.class))).thenReturn(Optional.empty());
         assertThat(web3SecretService.getDecryptedValue(secretAddress)).isEmpty();
-        verify(web3SecretRepository, times(1)).findById(any(Web3SecretHeader.class));
         verifyNoInteractions(encryptionService);
     }
     // endregion
@@ -151,19 +173,18 @@ class Web3SecretServiceTests {
     @Test
     void shouldGetEncryptedSecret() {
         Web3Secret encryptedSecret = new Web3Secret(secretAddress, encryptedSecretValue);
-        when(web3SecretRepository.findById(any(Web3SecretHeader.class))).thenReturn(Optional.of(encryptedSecret));
+        web3SecretRepository.save(encryptedSecret);
         Optional<Web3Secret> result = web3SecretService.getSecret(secretAddress);
         assertThat(result)
-                .contains(encryptedSecret);
-        verify(web3SecretRepository, times(1)).findById(any(Web3SecretHeader.class));
+                .isNotEmpty()
+                .usingRecursiveComparison()
+                .isEqualTo(Optional.of(encryptedSecret));
         verifyNoInteractions(encryptionService);
     }
 
     @Test
     void shouldGetEmptySecretIfSecretNotPresent() {
-        when(web3SecretRepository.findById(any(Web3SecretHeader.class))).thenReturn(Optional.empty());
         assertThat(web3SecretService.getSecret(secretAddress)).isEmpty();
-        verify(web3SecretRepository, times(1)).findById(any(Web3SecretHeader.class));
         verifyNoInteractions(encryptionService);
     }
     // endregion