+++ CHANGELOG.md
@@ -13,12 +13,14 @@ All notable changes to this project will be documented in this file.
 
 - Validate authorization first when working on compute secrets. (#229)
 - Remove outdated empty log assertions in session tests. (#233)
+- Do not use `@Data` lombok annotations on entities. (#235)
 
 ### Quality
 
 - Optimize chain calls in `AuthorizationService`. (#228)
 - Fix licence headers in some source files. (#232)
 - Move `EthereumCredentials` class to `com.iexec.sms.tee.challenge` package. (#234)
+- Use `@DataJpaTest` to run secrets and TEE challenges tests on H2 database. (#236 #237)
 
 ## [[8.4.0]](https://github.com/iExecBlockchainComputing/iexec-sms/releases/tag/v8.4.0) 2024-01-10